nvim: split additional struct types to types.go
